update/ adjust PR comment: rename checkIsFollowed
