+++ common/helpers/checkIsFollowed.tsx
